+++ lib/src/bones_ui_layout.dart
@@ -865,7 +865,7 @@ class UILayout {
     if (evaluated is Element) {
       return valueFromElement(evaluated);
     }
-    return '${evaluated}';
+    return '$evaluated';
   }
 
   Map<String, int> _buildEvaluationContext() {